+++ shell/edit/provisioning.cattle.io.cluster/Basics.vue
